modularize [file..] data & other small stuff
- modularize [file..] data, when loading a file now, only the size stat and the mimeType is given.
this is so the page can load intantly. Afterwards, a request for the
file content is made, and lastly if needed, a syntax highlighting

these subsequent requests also need to be handled if a user changes
files when the request is not finished, so an `AbortController` is
needed, so that store was made.

- for code files (.js, .ts, .py), I now use just shiki as it seems
  a lot faster than shiki-twoslash.

- use encodeURIComponent rather than replacing '/' with '\'