fix concurrent active list items, fix src path in rendered markdown, fix audio/video elements
concurrent active list items: basically when in dev mode, if you change
like the text content of an element, vite just hot fixes it and doesn't
like re run everything. This means the previous a element that was
cached to activeTag was an old element that is no longer in the dom

to fix this I simply removed the cached activeTag variable, and moved it
into switchActive

fix src path in rendered markdown: when converting markdown to html, the
program also fixes the src path on images. This however only worked if
the LFV_DEFAULT_FOLDER wasn't an absolute path or didn't have any back
tracking (../folder/other-folder)

fix audio/video elements: if you are previewing a video/audio element,
if you go to another page of that respective type, the src of the source
element will change, but the element will still show the same content as
it's not re-rendered

I simply used {#key} to re-render on servePath change